Flatten battery stats data layer
