+++ batterysync/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/data/batterystats/BatteryStatsRepositoryImpl.kt
@@ -1,23 +1,32 @@
-package com.boswelja.smartwatchextensions.batterysync.data.local
+package com.boswelja.smartwatchextensions.batterysync.data.batterystats
 
+import android.content.Context
 import com.boswelja.smartwatchextensions.batterysync.BatteryStats
+import com.boswelja.smartwatchextensions.batterysync.batteryStats
 import com.boswelja.smartwatchextensions.batterysync.database.BatteryStatsDatabase
+import com.boswelja.smartwatchextensions.batterysync.domain.repository.BatteryStatsRepository
 import com.squareup.sqldelight.runtime.coroutines.asFlow
 import com.squareup.sqldelight.runtime.coroutines.mapToOneOrNull
+import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.withContext
-import kotlin.coroutines.CoroutineContext
 
-class WatchBatteryStatsDbDataSource(
+/**
+ * An implementation of [BatteryStatsRepository] combining potentially multiple data sources.
+ */
+class BatteryStatsRepositoryImpl(
+    private val context: Context,
     private val database: BatteryStatsDatabase,
-    private val dispatcher: CoroutineContext
-) {
+    private val dispatcher: CoroutineDispatcher
+) : BatteryStatsRepository {
 
-    /**
-     * Flow the battery stats for a watch with a given ID.
-     * @param watchId The ID of the watch to flow stats for.
-     */
-    fun batteryStatsFor(watchId: String): Flow<BatteryStats?> = database
+    override fun getBatteryStatsForPhone(): BatteryStats? {
+        return context.batteryStats()
+    }
+
+    override fun getBatteryStatsForWatch(
+        watchId: String
+    ): Flow<BatteryStats?> = database
         .watchBatteryStatsQueries
         .getFor(watchId) { _, percent, charging, timestamp ->
             BatteryStats(percent, charging, timestamp)
@@ -25,22 +34,16 @@ class WatchBatteryStatsDbDataSource(
         .asFlow()
         .mapToOneOrNull()
 
-    /**
-     * Remove the battery stats for a watch with the given ID.
-     * @param watchId The ID of the watch whose battery stats should be removed.
-     */
-    suspend fun removeStatsFor(watchId: String) {
+    override suspend fun deleteBatteryStatsForWatch(watchId: String) {
         withContext(dispatcher) {
             database.watchBatteryStatsQueries.removeFor(watchId)
         }
     }
 
-    /**
-     * Update the battery stats for a watch with the given ID.
-     * @param watchId The ID of the watch whose stats should be updated.
-     * @param newStats The new [BatteryStats] to store.
-     */
-    suspend fun updateStatsFor(watchId: String, newStats: BatteryStats) {
+    override suspend fun putBatteryStatsForWatch(
+        watchId: String,
+        newStats: BatteryStats
+    ) {
         withContext(dispatcher) {
             database.watchBatteryStatsQueries.insert(
                 watchId,
